+++ include/BiuLabConfig.hh
@@ -1,7 +1,7 @@
 #ifndef BIULAB_CONFIG_H_
 #define BIULAB_CONFIG_H_
 
-#define BIULAB_VERSION "v1.1.0"
+#define BIULAB_VERSION "v1.1.1"
 #define BIULAB_INSTALL_PATH "/home/elonkou/ELONKOU/03.BIULAB/biulab/install"
 #define BIULAB_MODULES_PATH "/home/elonkou/ELONKOU/03.BIULAB/biulab/install/bin"
 #define BIULAB_APPLICATION_PATH "/home/elonkou/ELONKOU/03.BIULAB/biulab/install/../app"